libNX HID initialization
What was added:
* HID Service
* Support for Mutexes and Conditional Variables
What was improved:
* Service API now creates one instance per Session rather than a single instance for all Sessions
* Changed std::map objects into std::unordered_map in KProcess
* Comments on enumeration values

+++ app/src/main/cpp/skyline/nce.h
@@ -112,7 +112,7 @@ namespace skyline {
          * @param pid The PID of the process (Defaults to currPid)
          * @return The value in the register
          */
-        u64 GetRegister(Wreg regId, pid_t pid = 0);
+        u32 GetRegister(Wreg regId, pid_t pid = 0);
 
         /**
          * @brief Set the value of a Wn register
